requirements.txt
influxdb-client
python-dotenv
python-dateutil